squash this align from Fedora

+++ rust.spec
@@ -1,6 +1,5 @@
-# Combined rust / cross-{arch}-rust) specfile
-Name: rust
-# Keep Name on top !
+# Only x86_64 and i686 are Tier 1 platforms at this time.
+# https://forge.rust-lang.org/platform-support.html
 
 %global rustc_version 1.44.0
 %global cargo_version 1.44.0
@@ -11,12 +10,19 @@ Name: rust
 %global rust_triple armv7-unknown-linux-gnueabihf
 %endif
 
+%global python python3
+
+# LLDB isn't available everywhere...
+%bcond_without lldb
+
+Name:           rust
 Version:        %{rustc_version}
 Release:        1
 Summary:        The Rust Programming Language
 License:        (ASL 2.0 or MIT) and (BSD and MIT)
+# ^ written as: (rust itself) and (bundled libraries)
 URL:            https://www.rust-lang.org
-# ExclusiveArch:  %{ix86}
+
 
 %global rustc_package rustc-%{rustc_version}-src
 Source0:        https://static.rust-lang.org/dist/rustc-%{rustc_version}-src.tar.gz
@@ -28,6 +34,9 @@ Patch1: 0001-Use-a-non-existent-test-path-instead-of-clobbering-d.patch
 %global local_rust_root %{_builddir}/%{bootstrap_root}/usr
 %global bootstrap_source rust-%{rustc_version}-%{rust_triple}.tar.gz
 
+BuildRequires:  make
+BuildRequires:  cmake
+BuildRequires:  gcc
 BuildRequires:  gcc-c++
 BuildRequires:  ncurses-devel
 BuildRequires:  curl
@@ -35,11 +44,27 @@ BuildRequires:  pkgconfig(libcurl)
 BuildRequires:  pkgconfig(liblzma)
 BuildRequires:  pkgconfig(openssl)
 BuildRequires:  pkgconfig(zlib)
-BuildRequires:  python
+BuildRequires:  python3-base
 BuildRequires:  llvm-devel
 BuildRequires:  libffi-devel
 
-Requires:       %{name}-std-static = %{rustc_version}-%{release}
+# make check needs "ps" for src/test/run-pass/wait-forked-but-failed-child.rs
+BuildRequires:  procps
+
+# debuginfo-gdb tests need gdb
+BuildRequires:  gdb
+
+
+# Virtual provides for folks who attempt "dnf install rustc"
+Provides:       rustc = %{version}-%{release}
+Provides:       rustc%{?_isa} = %{version}-%{release}
+
+# Always require our exact standard library
+Requires:       %{name}-std-static%{?_isa} = %{version}-%{release}
+
+# The C compiler is needed at runtime just for linking.  Someday rustc might
+# invoke the linker directly, and then we'll only need binutils.
+# https://github.com/rust-lang/rust/issues/11937
 Requires:       /usr/bin/cc
 
 # ALL Rust libraries are private, because they don't keep an ABI.
@@ -85,32 +110,118 @@ This package includes the common functionality for %{name}-gdb and %{name}-lldb.
 Summary:        GDB pretty printers for Rust
 BuildArch:      noarch
 Requires:       gdb
-Requires:       %{name}-debugger-common = %{rustc_version}-%{release}
+Requires:       %{name}-debugger-common = %{version}-%{release}
 
 %description gdb
 This package includes the rust-gdb script, which allows easier debugging of Rust
 programs.
 
 
+%if %with lldb
+
 %package lldb
 Summary:        LLDB pretty printers for Rust
-
+BuildArch:      noarch
 Requires:       lldb
-Requires:       python2-lldb
-Requires:       %{name}-debugger-common = %{rustc_version}-%{release}
+Requires:       python3-lldb
+Requires:       %{name}-debugger-common = %{version}-%{release}
 
 %description lldb
 This package includes the rust-lldb script, which allows easier debugging of Rust
 programs.
 
+%endif
+
+
+%package doc
+Summary:        Documentation for Rust
+# NOT BuildArch:      noarch
+# Note, while docs are mostly noarch, some things do vary by target_arch.
+# Koji will fail the build in rpmdiff if two architectures build a noarch
+# subpackage differently, so instead we have to keep its arch.
+
+%description doc
+This package includes HTML documentation for the Rust programming language and
+its standard library.
+
 
 %package -n cargo
 Summary:        Rust's package manager and build tool
 Version:        %{cargo_version}
+# Cargo is not much use without Rust
+Requires:       rust
+# For tests:
+BuildRequires:  git
 
 %description -n cargo
 Cargo is a tool that allows Rust projects to declare their various dependencies
 and ensure that you'll always get a repeatable build.
+
+
+%package -n cargo-doc
+Summary:        Documentation for Cargo
+BuildArch:      noarch
+# Cargo no longer builds its own documentation
+# https://github.com/rust-lang/cargo/pull/4904
+Requires:       rust-doc = %{version}-%{release}
+
+%description -n cargo-doc
+This package includes HTML documentation for Cargo.
+
+
+%package -n rustfmt
+Summary:        Tool to find and fix Rust formatting issues
+Requires:       cargo
+
+%description -n rustfmt
+A tool for formatting Rust code according to style guidelines.
+
+
+%package -n rls
+Summary:        Rust Language Server for IDE integration
+Requires:       rust-analysis
+# /usr/bin/rls is dynamically linked against internal rustc libs
+Requires:       %{name}%{?_isa} = %{version}-%{release}
+
+%description -n rls
+The Rust Language Server provides a server that runs in the background,
+providing IDEs, editors, and other tools with information about Rust programs.
+It supports functionality such as 'goto definition', symbol search,
+reformatting, and code completion, and enables renaming and refactorings.
+
+
+%package -n clippy
+Summary:        Lints to catch common mistakes and improve your Rust code
+Requires:       cargo
+# /usr/bin/clippy-driver is dynamically linked against internal rustc libs
+Requires:       %{name}%{?_isa} = %{version}-%{release}
+
+# The component/package was clippy-preview until Rust 1.31.
+Obsoletes:      clippy-preview <= 0.0.212
+Provides:       clippy-preview = %{version}-%{release}
+
+%description -n clippy
+A collection of lints to catch common mistakes and improve your Rust code.
+
+
+%package src
+Summary:        Sources for the Rust standard library
+BuildArch:      noarch
+
+%description src
+This package includes source files for the Rust standard library.  It may be
+useful as a reference for code completion tools in various editors.
+
+
+%package analysis
+Summary:        Compiler analysis data for the Rust standard library
+Requires:       rust-std-static%{?_isa} = %{version}-%{release}
+
+%description analysis
+This package contains analysis data files produced with rustc's -Zsave-analysis
+feature for the Rust standard library. The RLS (Rust Language Server) uses this
+data to provide information about the Rust standard library.
+
 
 %prep
 
@@ -121,15 +232,42 @@ test -f '%{local_rust_root}/bin/cargo'
 test -f '%{local_rust_root}/bin/rustc'
 
 %setup -q -n %{rustc_package}
+
 %patch1 -p1
 
-rm -rf src/llvm/
-rm -rf src/llvm-emscripten/
+sed -i.try-py3 -e '/try python2.7/i try python3 "$@"' ./configure
+
+rm -rf src/llvm-project/
 
 # We never enable other LLVM tools.
 rm -rf src/tools/clang
 rm -rf src/tools/lld
 rm -rf src/tools/lldb
+
+# Remove other unused vendored libraries
+rm -rf vendor/curl-sys/curl/
+rm -rf vendor/jemalloc-sys/jemalloc/
+rm -rf vendor/libz-sys/src/zlib/
+rm -rf vendor/lzma-sys/xz-*/
+rm -rf vendor/openssl-src/openssl/
+
+	# rename bundled license for packaging
+cp -a vendor/backtrace-sys/src/libbacktrace/LICENSE{,-libbacktrace}
+
+# Static linking to distro LLVM needs to add -lffi
+# https://github.com/rust-lang/rust/issues/34486
+sed -i.ffi -e '$a #[link(name = "ffi")] extern {}' \
+  src/librustc_llvm/lib.rs
+
+# The configure macro will modify some autoconf-related files, which upsets
+# cargo when it tries to verify checksums in those files.  If we just truncate
+# that file list, cargo won't have anything to complain about.
+find vendor -name .cargo-checksum.json \
+  -exec sed -i.uncheck -e 's/"files":{[^}]*}/"files":{ }/' '{}' '+'
+
+# Sometimes Rust sources start with #![...] attributes, and "smart" editors think
+# it's a shebang and make them executable. Then brp-mangle-shebangs gets upset...
+find -name '*.rs' -type f -perm /111 -exec chmod -v -x '{}' '+'
 
 
 %build
@@ -143,32 +281,36 @@ export RUSTFLAGS="%{rustflags}"
 
 # full debuginfo is exhausting memory; just do libstd for now
 # https://github.com/rust-lang/rust/issues/45854
-%define enable_debuginfo --disable-debuginfo --disable-debuginfo-only-std --disable-debuginfo-tools --disable-debuginfo-lines
+%define enable_debuginfo --debuginfo-level=0 --debuginfo-level-std=2
+# %define enable_debuginfo --disable-debuginfo --disable-debuginfo-only-std --disable-debuginfo-tools --disable-debuginfo-lines
+
+%define codegen_units_std --set rust.codegen-units-std=1
 
 %configure --disable-option-checking \
   --libdir=%{common_libdir} \
   --build=%{rust_triple} --host=%{rust_triple} --target=%{rust_triple} \
+  --python=%{python} \
   --local-rust-root=%{local_rust_root} \
   --enable-local-rebuild \
   --enable-llvm-link-shared \
   --enable-optimize \
   --disable-rpath \
-  --disable-docs \
-  --disable-compiler-docs \
   --disable-codegen-tests \
   %{enable_debuginfo} \
-  --enable-vendor \
   --enable-extended \
+  --enable-vendor \
+  %{?codegen_units_std} \
   --tools=cargo \
   --llvm-root=/usr/
 
-python ./x.py build
+%{python} ./x.py build
+%{python} ./x.py doc
 
 
 %install
 export RUSTFLAGS="%{rustflags}"
 
-DESTDIR=%{buildroot} python ./x.py install
+DESTDIR=%{buildroot} %{python} ./x.py install
 
 # Make sure the shared libraries are in the proper libdir
 %if "%{_libdir}" != "%{common_libdir}"
@@ -187,9 +329,11 @@ find %{buildroot}%{_libdir} -maxdepth 1 -type f -name '*.so' \
 (cd "%{buildroot}%{rustlibdir}/%{rust_triple}/lib" &&
  find ../../../../%{_lib} -maxdepth 1 -name '*.so' |
  while read lib; do
-   # make sure they're actually identical!
-   cmp "$lib" "${lib##*/}"
-   ln -v -f -s -t . "$lib"
+   if [ -f "${lib##*/}" ]; then
+     # make sure they're actually identical!
+     cmp "$lib" "${lib##*/}"
+     ln -v -f -s -t . "$lib"
+   fi
  done)
 
 # Remove installer artifacts (manifests, uninstall scripts, etc.)
@@ -198,7 +342,7 @@ find %{buildroot}%{rustlibdir} -maxdepth 1 -type f -exec rm -v '{}' '+'
 # Remove backup files from %%configure munging
 find %{buildroot}%{rustlibdir} -type f -name '*.orig' -exec rm -v '{}' '+'
 
-# Remove unwanted documentation files
+# Remove unwanted documentation files (we already package them)
 rm -f %{buildroot}%{_docdir}/%{name}/README.md
 rm -f %{buildroot}%{_docdir}/%{name}/COPYRIGHT
 rm -f %{buildroot}%{_docdir}/%{name}/LICENSE
@@ -207,11 +351,42 @@ rm -f %{buildroot}%{_docdir}/%{name}/LICENSE-MIT
 rm -f %{buildroot}%{_docdir}/%{name}/LICENSE-THIRD-PARTY
 rm -f %{buildroot}%{_docdir}/%{name}/*.old
 
+# Sanitize the HTML documentation
+find %{buildroot}%{_docdir}/%{name}/html -empty -delete
+find %{buildroot}%{_docdir}/%{name}/html -type f -exec chmod -x '{}' '+'
+
 # Create the path for crate-devel packages
 mkdir -p %{buildroot}%{_datadir}/cargo/registry
 
+# Cargo no longer builds its own documentation
+# https://github.com/rust-lang/cargo/pull/4904
+mkdir -p %{buildroot}%{_docdir}/cargo
+ln -sT ../rust/html/cargo/ %{buildroot}%{_docdir}/cargo/html
+
+%if %without lldb
+rm -f %{buildroot}%{_bindir}/rust-lldb
+rm -f %{buildroot}%{rustlibdir}/etc/lldb_*.py*
+%endif
+
+
+%check
+%{?cmake_path:export PATH=%{cmake_path}:$PATH}
+%{?rustflags:export RUSTFLAGS="%{rustflags}"}
+
+# The results are not stable on koji, so mask errors and just log it.
+%{python} ./x.py test --no-fail-fast || :
+%{python} ./x.py test --no-fail-fast cargo || :
+%{python} ./x.py test --no-fail-fast clippy || :
+%{python} ./x.py test --no-fail-fast rls || :
+%{python} ./x.py test --no-fail-fast rustfmt || :
+
+
+%ldconfig_scriptlets
+
+
 %files
 %license COPYRIGHT LICENSE-APACHE LICENSE-MIT
+%license vendor/backtrace-sys/src/libbacktrace/LICENSE-libbacktrace
 %doc README.md
 %{_bindir}/rustc
 %{_bindir}/rustdoc
@@ -222,7 +397,7 @@ mkdir -p %{buildroot}%{_datadir}/cargo/registry
 %dir %{rustlibdir}/%{rust_triple}
 %dir %{rustlibdir}/%{rust_triple}/lib
 %{rustlibdir}/%{rust_triple}/lib/*.so
-%{rustlibdir}/%{rust_triple}/codegen-backends/
+%exclude %{_bindir}/*miri
 
 
 %files std-static
@@ -240,13 +415,31 @@ mkdir -p %{buildroot}%{_datadir}/cargo/registry
 
 %files gdb
 %{_bindir}/rust-gdb
-%{_bindir}/rust-gdbgui
 %{rustlibdir}/etc/gdb_*.py*
+%exclude %{_bindir}/rust-gdbgui
 
 
+%if %with lldb
 %files lldb
 %{_bindir}/rust-lldb
 %{rustlibdir}/etc/lldb_*.py*
+%endif
+
+
+%files doc
+%docdir %{_docdir}/%{name}
+%dir %{_docdir}/%{name}
+%dir %{_docdir}/%{name}/html
+%{_docdir}/%{name}/html/*/
+%{_docdir}/%{name}/html/*.html
+%{_docdir}/%{name}/html/*.css
+%{_docdir}/%{name}/html/*.ico
+%{_docdir}/%{name}/html/*.js
+%{_docdir}/%{name}/html/*.png
+%{_docdir}/%{name}/html/*.svg
+%{_docdir}/%{name}/html/*.woff
+%license %{_docdir}/%{name}/html/*.txt
+%license %{_docdir}/%{name}/html/*.md
 
 
 %files -n cargo
@@ -259,17 +452,43 @@ mkdir -p %{buildroot}%{_datadir}/cargo/registry
 %dir %{_datadir}/cargo
 %dir %{_datadir}/cargo/registry
 
+
+%files -n cargo-doc
+%docdir %{_docdir}/cargo
+%dir %{_docdir}/cargo
+%{_docdir}/cargo/html
+
+
+%files -n rustfmt
+%{_bindir}/rustfmt
+%{_bindir}/cargo-fmt
+%doc src/tools/rustfmt/{README,CHANGELOG,Configurations}.md
+%license src/tools/rustfmt/LICENSE-{APACHE,MIT}
+
+
+%files -n rls
+%{_bindir}/rls
+%doc src/tools/rls/{README.md,COPYRIGHT,debugging.md}
+%license src/tools/rls/LICENSE-{APACHE,MIT}
+
+
+%files -n clippy
+%{_bindir}/cargo-clippy
+%{_bindir}/clippy-driver
+%doc src/tools/clippy/{README.md,CHANGELOG.md}
+%license src/tools/clippy/LICENSE-{APACHE,MIT}
+
+
+%files src
+%dir %{rustlibdir}
+%{rustlibdir}/src
+
+
+%files analysis
+%{rustlibdir}/%{rust_triple}/analysis/
+
+
 %changelog
-* Fri Apr 12 2019 Lucien Xu <sfietkonstantin@free.fr> - 1.34.0-1
-- Package 1.34.0
-
-* Sat Mar 31 2019 Lucien Xu <sfietkonstantin@free.fr> - 1.33.0-1
-- Package 1.33.0
-
-* Sat Feb 09 2019 Lucien Xu <sfietkonstantin@free.fr> - 1.32.0-1
-- Package 1.32.0
-
-* Sat Dec 08 2018 Lucien Xu <sfietkonstantin@free.fr> - 1.31.0-1
-- Package 1.31.0
-- Based on Fedora packaging
+* Thu Jun 12 2020 Raine Makelainen <raine.makelainen@jolla.com> - 1.44.0-1
+- Integrate 1.44.0.
